requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pyyaml>=6.0
numpy>=1.26
scipy>=1.11
pandas>=2.2
pyarrow>=15.0
pulp>=2.8
jinja2>=3.1
httpx>=0.27
pytest>=8.0
pytest-cov>=5.0
pytest-asyncio>=0.23
